backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
anthropic>=0.52.0
pydantic>=2.10.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pypdf>=4.2.0
python-docx>=1.1.0
